+++ src/compas_blender/artists/frameartist.py
@@ -1,3 +1,9 @@
+import bpy
+from typing import List
+from typing import Optional
+
+from compas.geometry import Frame
+
 import compas_blender
 from .artist import BlenderArtist
 
@@ -52,8 +58,11 @@ class FrameArtist(BlenderArtist):
             artist.draw()
 
     """
-    def __init__(self, frame, collection=None, scale=1.0):
-        super().__init__()
+    def __init__(self,
+                 frame: Frame,
+                 collection: Optional[bpy.types.Collection] = None,
+                 scale: float = 1.0):
+        super(FrameArtist, self).__init__()
         self.collection = collection
         self.frame = frame
         self.scale = scale or 1.0
@@ -62,7 +71,7 @@ class FrameArtist(BlenderArtist):
         self.color_yaxis = (0, 255, 0)
         self.color_zaxis = (0, 0, 255)
 
-    def draw(self):
+    def draw(self) -> None:
         """Draw the frame.
 
         Returns
@@ -73,7 +82,7 @@ class FrameArtist(BlenderArtist):
         self.draw_origin()
         self.draw_axes()
 
-    def draw_origin(self):
+    def draw_origin(self) -> List[bpy.types.Object]:
         """Draw the origin of the frame.
 
         Returns
@@ -90,7 +99,7 @@ class FrameArtist(BlenderArtist):
         self.objects += objects
         return objects
 
-    def draw_axes(self):
+    def draw_axes(self) -> List[bpy.types.Object]:
         """Draw the axes of the frame.
 
         Returns